+++ index.js
@@ -1,4 +1,6 @@
 const EventEmitter = require( "events" ),
+      AppSwapStrategy = require( "./Lib/Strategy/AppSwap" ),
+      ScriptSwapStrategy = require( "./Lib/Strategy/ScriptSwap" ),
       semver = require( "semver" ),
       os = require( "os" ),
       { join, basename, dirname } = require( "path" ),
@@ -9,9 +11,8 @@ const EventEmitter = require( "events" ),
       { readJson, download }  = require( "./Lib/request" ),
       { launch, rtrim } = require( "./Lib/utils" ),
       { PLATFORM_FULL, swapFactory,
-        getExecutable, UPDATE_DIR, EXEC_DIR, BACKUP_DIR, LOG_DIR } = require( "./Lib/env" ),
+        getExecutable, UPDATE_DIR, EXEC_DIR, BACKUP_DIR, LOG_PATH } = require( "./Lib/env" ),
 
-      LOG_FILE = "nw-autoupdater.log",
       ERR_INVALID_REMOTE_MANIFEST = "Invalid manifest structure",
       DEBOUNCE_TIME = 500,
 
@@ -20,10 +21,12 @@ const EventEmitter = require( "events" ),
         backupDir: BACKUP_DIR,
         execDir: EXEC_DIR,
         updateDir: UPDATE_DIR,
-        logDir: LOG_DIR,
+        logPath: LOG_PATH,
         verbose: false,
-        swapScript: null
+        swapScript: null,
+        strategy: "AppSwap"
       };
+
 
 class AutoUpdater extends EventEmitter {
   /**
@@ -44,8 +47,11 @@ class AutoUpdater extends EventEmitter {
     this.argv = nw.App.argv;
     this.remoteManifest = "";
     this.options = Object.assign( {}, DEFAULT_OPTIONS, options );
+
     this.options.execDir = rtrim( this.options.execDir );
     this.options.executable = this.options.executable || getExecutable( manifest.name );
+
+    Object.assign( this, this.options.strategy === "ScriptSwap" ? ScriptSwapStrategy : AppSwapStrategy );
 
   }
 
@@ -54,7 +60,11 @@ class AutoUpdater extends EventEmitter {
    * @returns {Promise<JSON>}
    */
   async readRemoteManifest(){
-    return await readJson( this.manifest.manifestUrl );
+    try {
+      return await readJson( this.manifest.manifestUrl );
+    } catch ( e ) {
+      throw new Error( `Cannot read remote manifest from ${this.manifest.manifestUrl}` );
+    }
   }
   /**
    * Check if a new app version available
@@ -84,7 +94,11 @@ class AutoUpdater extends EventEmitter {
     const onProgress = ( length ) => {
       this.emit( "download", length, release.size );
     };
-    return await download( release.url, os.tmpdir(), debounce( onProgress, debounceTime ));
+    try {
+      return await download( release.url, os.tmpdir(), debounce( onProgress, debounceTime ));
+    } catch ( e ) {
+      throw new Error( `Cannot download package from ${release.url}` );
+    }
   }
   /**
    * Unpack downloaded version
@@ -106,10 +120,18 @@ class AutoUpdater extends EventEmitter {
 
     switch( true ) {
       case isGzRe.test( updateFile ):
-         await unpackTarGz( updateFile, updateDir, debounce( onProgress, debounceTime ) );
+         try {
+          await unpackTarGz( updateFile, updateDir, debounce( onProgress, debounceTime ) );
+         } catch ( e ) {
+            throw new Error( `Cannot unpack .tar.gz package ${updateFile}` );
+         }
          break;
       case isZipRe.test( updateFile ):
-         await unpackZip( updateFile, updateDir, debounce( onProgress, debounceTime ) );
+         try {
+          await unpackZip( updateFile, updateDir, debounce( onProgress, debounceTime ) );
+         } catch ( e ) {
+            throw new Error( `Cannot unpack .zip package ${updateFile}` );
+         }
          break;
       default:
          throw new Error( "Release arhive of unsuported type" );
@@ -118,20 +140,6 @@ class AutoUpdater extends EventEmitter {
     return updateDir;
   }
 
-  /**
-   * Restart and launch detached swap
-   * @returns {Promise}
-   */
-  async restartToSwap(){
-    const { updateDir, logDir  } = this.options,
-          swap = swapFactory( this.options ),
-          logPath = join( logDir, LOG_FILE ),
-          args = swap.getArgs();
-
-    swap.extractScript( updateDir );
-    await launch( swap.getRunner(), args, updateDir, logPath );
-    nw.App.quit();
-  }
   /**
    * @deprecated since v.1.1.0
    * @returns {Boolean}
@@ -154,5 +162,163 @@ class AutoUpdater extends EventEmitter {
     return false;
   }
 }
+
+
+////const EventEmitter = require( "events" ),
+//      semver = require( "semver" ),
+//      os = require( "os" ),
+//      { join, basename, dirname } = require( "path" ),
+//      unpackTarGz = require( "./Lib/unpackTarGz" ),
+//      unpackZip = require( "./Lib/unpackZip" ),
+//      debounce = require( "debounce" ),
+//
+//      { readJson, download }  = require( "./Lib/request" ),
+//      { launch, rtrim } = require( "./Lib/utils" ),
+//      { PLATFORM_FULL, swapFactory,
+//        getExecutable, UPDATE_DIR, EXEC_DIR, BACKUP_DIR, LOG_DIR } = require( "./Lib/env" ),
+//
+//      LOG_FILE = "nw-autoupdater.log",
+//      ERR_INVALID_REMOTE_MANIFEST = "Invalid manifest structure",
+//      DEBOUNCE_TIME = 500,
+//
+//      DEFAULT_OPTIONS = {
+//        executable: null,
+//        backupDir: BACKUP_DIR,
+//        execDir: EXEC_DIR,
+//        updateDir: UPDATE_DIR,
+//        logDir: LOG_DIR,
+//        verbose: false,
+//        swapScript: null
+//      };
+//
+//class AutoUpdater extends EventEmitter {
+//  /**
+//   * Create AutoUpdate
+//   * @param {Object} manifest
+//   * @param {Object} options
+//   */
+//  constructor( manifest, options = {}){
+//
+//    super();
+//
+//    this.manifest = manifest;
+//    if ( !this.manifest.manifestUrl ) {
+//      throw new Error( `Manifest must contain manifestUrl field` );
+//    }
+//
+//    this.release = "";
+//    this.argv = nw.App.argv;
+//    this.remoteManifest = "";
+//    this.options = Object.assign( {}, DEFAULT_OPTIONS, options );
+//    this.options.execDir = rtrim( this.options.execDir );
+//    this.options.executable = this.options.executable || getExecutable( manifest.name );
+//
+//  }
+//
+//  /**
+//   * Read package.json from the release server
+//   * @returns {Promise<JSON>}
+//   */
+//  async readRemoteManifest(){
+//    return await readJson( this.manifest.manifestUrl );
+//  }
+//  /**
+//   * Check if a new app version available
+//   * @param {Object} remoteManifest
+//   * @returns {Promise<boolean>}
+//   */
+//  async checkNewVersion( remoteManifest ){
+//    if ( !remoteManifest || !remoteManifest.packages ){
+//      throw new TypeError( ERR_INVALID_REMOTE_MANIFEST );
+//    }
+//    return semver.gt( remoteManifest.version, this.manifest.version );
+//  }
+//  /**
+//   * Download new version
+//   * @param {Object} remoteManifest
+//   * @param {Object} options
+//   * @returns {Promise<string>}
+//   */
+//  async download( remoteManifest, { debounceTime } = { debounceTime: DEBOUNCE_TIME }){
+//    if ( !remoteManifest || !remoteManifest.packages ){
+//      throw new TypeError( ERR_INVALID_REMOTE_MANIFEST );
+//    }
+//    const release = remoteManifest.packages[ PLATFORM_FULL ];
+//    if ( !release ) {
+//      throw new Error( `No release matches the platfrom ${PLATFORM_FULL}` );
+//    }
+//    const onProgress = ( length ) => {
+//      this.emit( "download", length, release.size );
+//    };
+//    return await download( release.url, os.tmpdir(), debounce( onProgress, debounceTime ));
+//  }
+//  /**
+//   * Unpack downloaded version
+//   * @param {string} updateFile
+//   * @param {Object} options
+//   * @returns {Promise<string>}
+//   */
+//  async unpack( updateFile, { debounceTime } = { debounceTime: DEBOUNCE_TIME } ){
+//    const isZipRe = /\.zip$/i,
+//          isGzRe = /\.tar\.gz$/i,
+//          onProgress = ( installFiles, totalFiles ) => {
+//            this.emit( "install", installFiles, totalFiles );
+//          },
+//          updateDir = this.options.updateDir;
+//
+//    if ( !updateFile ){
+//      throw new Error( "You have to call first download method" );
+//    }
+//
+//    switch( true ) {
+//      case isGzRe.test( updateFile ):
+//         await unpackTarGz( updateFile, updateDir, debounce( onProgress, debounceTime ) );
+//         break;
+//      case isZipRe.test( updateFile ):
+//         await unpackZip( updateFile, updateDir, debounce( onProgress, debounceTime ) );
+//         break;
+//      default:
+//         throw new Error( "Release arhive of unsuported type" );
+//         break;
+//    }
+//    return updateDir;
+//  }
+//
+//  /**
+//   * Restart and launch detached swap
+//   * @returns {Promise}
+//   */
+//  async restartToSwap(){
+//    const { updateDir, logDir  } = this.options,
+//          swap = swapFactory( this.options ),
+//          logPath = join( logDir, LOG_FILE ),
+//          args = swap.getArgs();
+//
+//    swap.extractScript( updateDir );
+//    await launch( swap.getRunner(), args, updateDir, logPath );
+//    nw.App.quit();
+//  }
+//  /**
+//   * @deprecated since v.1.1.0
+//   * @returns {Boolean}
+//   */
+//  isSwapRequest(){
+//    return false;
+//  }
+//  /**
+//   * @deprecated since v.1.1.0
+//   * @returns {Boolean}
+//   */
+//  async swap(){
+//    return false;
+//  }
+//  /**
+//   * @deprecated since v.1.1.0
+//   * @returns {Boolean}
+//   */
+//  async restart(){
+//    return false;
+//  }
+//}
 
 module.exports = AutoUpdater;